+++ www/js/services/facebookPluginService.js
@@ -8,6 +8,8 @@ function facebookPluginService(platformService){
     vm.getProfilePicture = getProfilePicture;
     vm.getUserProfile = getUserProfile;
     vm.getUserLocation = getUserLocation;
+    vm.createUserAccount = createUserAccount;
+    vm.isExistingUserHasFacebookProp = isExistingUserHasFacebookProp;
 
     function login(successCallback){
 
@@ -23,15 +25,6 @@ function facebookPluginService(platformService){
     }
 
     function getProfilePicture(userID, successCallback){
-
-        /**
-         * API: https://developers.facebook.com/docs/graph-api/reference/user/picture/
-         * 
-         * By default this edge will return a 302 redirect to the picture image.
-         * To get access to the data about the picture, please include redirect=false in your query.
-         * 
-         *  */ 
-
         facebookConnectPlugin
             .api('/' + userID + '/picture?redirect=false', ["public_profile"], successCallback, 
             function errorCallback(error){
@@ -61,6 +54,71 @@ function facebookPluginService(platformService){
             function errorCallback(error){
                 throw error;
             });
+    }
+
+    function createUserAccount(callback){
+        vm.login(function(fbLoginData){
+            var fbAccessToken = fbLoginData.authResponse.accessToken,
+                fbUserID = fbLoginData.authResponse.userID;
+            vm.getUserProfile(function(fbUserProfile){
+                var data = createUserWithFacebookProp(fbUserID, fbAccessToken, fbUserProfile);
+
+                callback(data);
+            });
+        });
+    }
+
+    function createUserWithFacebookProp(fbUserID, fbAccessToken, fbUserProfile){
+
+        if(!fbUserID || !fbAccessToken || !fbUserProfile)
+            throw new Error('Failed to create User with Facebook Prop');
+        if(!fbUserProfile.hasOwnProperty('first_name') || !fbUserProfile.hasOwnProperty('last_name') 
+            || !fbUserProfile.hasOwnProperty('gender') || !fbUserProfile.hasOwnProperty('email'))
+            throw new Error('fbUserProfile missing props')
+
+        var data = {
+            user: {
+                email: fbUserProfile.email,
+                username: [fbUserProfile.first_name, fbUserProfile.last_name].join('').toLowerCase(),
+                password: fbUserID,
+                account: {
+                    profile: {
+                        firstname: fbUserProfile.first_name,
+                        lastname: fbUserProfile.last_name,
+                        gender: fbUserProfile.gender
+                    }
+                }
+            }
+        };
+        data.user.account.social = {
+            facebook: {
+                profile: data.user.account.profile,
+                accessToken: fbAccessToken,
+                userID: fbUserID
+            }
+        };
+
+        return data;
+    }
+
+    function isExistingUserHasFacebookProp(email){
+        return User.findOne({
+            filter: {
+                where: {
+                    email: email
+                }
+            }
+        }).$promise.then(function(value, responseHeaders){
+             if(!value.hasOwnProperty('error')){
+                if(value.account.hasOwnProperty('social')){
+                    if(value.account.social.hasOwnProperty('facebook')){
+                        return true;
+                    }else{
+                        return false;
+                    }
+                }
+            }
+        });
     }
 }