+++ www/js/services/loadingService.js
@@ -6,10 +6,20 @@ function loadingService($ionicLoading){
     var vm = this;
     vm.showLoading = showLoading;
     vm.hideLoading = hideLoading;
+    vm.successNotify = successNotify;
+    vm.errorNotify = errorNotify;
 
-    function showLoading(){
+    function showLoading(config){
+        var defaultConfig = {
+            template: config.template || 'Please wait... <ion-spinner></ion-spinner>',
+            noBackdrop: config.noBackdrop || false,
+            duration: config.duration || 0 
+        };
+
         return $ionicLoading.show({
-            template: 'Please wait... <ion-spinner></ion-spinner>'
+            template: defaultConfig.template,
+            noBackdrop: defaultConfig.noBackdrop,
+            duration: defaultConfig.duration
         }).then(function(){
             return 'LoadingService: showLoading';
         });
@@ -20,6 +30,26 @@ function loadingService($ionicLoading){
             .then(function(){
                 return 'LoadingService: hideLoading';
             });
+    }
+    
+    function successNotify(message, duration){
+        return vm.showLoading({
+            template: [message, '<i class="icon ion-checkmark balanced">'].join(' '),
+            noBackdrop: true,
+            duration: duration || 3000
+        }).then(function(){
+            return 'LoadingService: successNotify';
+        });
+    }
+
+    function errorNotify(message, duration){
+        return vm.showLoading({
+            template: [message, '<i class="icon ion-sad assertive">'].join(' '),
+            noBackdrop: true,
+            duration: duration || 3000
+        }).then(function(){
+            return 'LoadingService: successNotify';
+        });
     }
 }