+++ www/js/services/signupService.js
@@ -24,13 +24,12 @@ angular
 
             return User.create({}, data.user).$promise.then(function(value, responseHeaders){
                 if(value.hasOwnProperty('error')){
-                    if(value.error.status === 422 || value.error.statusCode === 422){
-                        throw new Error(JSON.stringify(value.error));
-                    }
+                    throw new Error(JSON.stringify(value.error));
                 }
                 return value.$promise;
             }).catch(function(err){
-                throw new Error('Stacktrace: ' + err);
+                err = JSON.parse(err.message);
+                throw new Error(err.message);
             });
         }
     }